update based on version and serial and mosltry sripts are working

+++ trunk/www/ewcp/cgi-bin/util_apps-update.cgi
@@ -20,15 +20,34 @@ fi
 
 
 cd $storage
-wget http://eboda-hd-for-all-500.googlecode.com/files/cb3pp-latest.zip
-[ $? == 0 ] || nice_exit 2
 
-rm -rf cb3pp/*
+SERIAL=0
+VERSION=v0.0
 
-unzip -o cb3pp-latest.zip
-rm cb3pp-latest.zip
 
-cp $storage/ewcp/S99ewcp /cb3pp/etc/init.d/S99ewcp
+wget http://eboda-hd-for-all-500.googlecode.com/files/cb3pp-latest-version.txt
+[ $? == 0 ] || nice_exit 1  
+. ./cb3pp-latest-version.txt
+. ./cb3pp/cb3pp-version.txt
+
+if [ $LATEST_SERIAL -gt $SERIAL ]
+then
+    echo "Latest version available is ${LATEST_VERSION}, you have $VERSION, updating !!!"
+
+
+    wget http://eboda-hd-for-all-500.googlecode.com/files/cb3pp-latest.zip
+    [ $? == 0 ] || nice_exit 2
+    
+    rm -rf cb3pp/*
+    
+    unzip -o cb3pp-latest.zip
+    rm cb3pp-latest.zip
+    
+    cp $storage/ewcp/S99ewcp /cb3pp/etc/init.d/S99ewcp
+    sh /cb3pp/etc/init.d/S99ewcp
+else
+    echo "You are already running the latest version ($VERSION)"
+fi
 
 
 nice_exit 0
